created useful names for contraints

+++ helpers.py
@@ -1,8 +1,9 @@
 # Print model
-def print_model(model):
+def print_model(model, delete_file=True):
     import os
     model.write("model.lp")
     with open("model.lp") as f:
         for line in f.readlines():
             print(line)
-    os.remove("model.lp")
+    if delete_file:
+        os.remove("model.lp")
